% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactmatrix.R
\name{expandC}
\alias{expandC}
\title{Expand the Contact Matrix over Regions}
\usage{
expandC(C, n)
}
\arguments{
\item{C}{a \code{\link{contactmatrix}}.}

\item{n}{the size of the secondary dimension to expand to.}
}
\value{
a square matrix with \code{n*ncol(C)} rows and columns.
}
\description{
This is simply the Kronecker product of the contact matrix \code{C}
with a matrix of ones of dimension \code{n} x \code{n}.
}
\examples{
expandC(contactmatrix(), 2)
}
