% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHHH4.R
\name{plotHHH4_fitted_groups}
\alias{plotHHH4_fitted_groups}
\title{Plot Mean Components of a \code{hhh4} Fit by Group}
\usage{
plotHHH4_fitted_groups(x, groups, total = FALSE, decompose = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{"hhh4"}.}

\item{groups}{a factor of grouping the units in the model, i.e., it must be
of length \code{x$nUnit}. There will be one plot for each factor level.}

\item{total}{a logical indicating if the group-wise mean components should
be subsequently summed up over all \code{groups} for an overall plot.}

\item{decompose, ...}{see \code{\link[surveillance]{plotHHH4_fitted}}.}
}
\value{
see \code{\link[surveillance]{plotHHH4_fitted}}.
}
\description{
Fitted mean components for age-structured, areal time series
\code{\link[surveillance]{hhh4}} models can be aggregated over districts or age groups.
}
