% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHHH4.R
\name{plotHHH4_maps_groups}
\alias{plotHHH4_maps_groups}
\title{Plot Mean Components of a \code{hhh4} Fit by District Averaged Over Time}
\usage{
plotHHH4_maps_groups(x, map, districts, ...)
}
\arguments{
\item{x}{an object of class \code{"hhh4"}.}

\item{map}{an object inheriting from \code{"SpatialPolygons"}.}

\item{districts}{a factor of length \code{x$nUnit} with as many levels as
there are districts and names according to \code{row.names(map)}.}

\item{...}{arguments passed to \code{\link{plotHHH4_maps}}.}
}
\value{
see \code{\link{plotHHH4_maps}}
}
\description{
This is a wrapper for \code{\link{plotHHH4_maps}} with prior aggregation
over different (age) groups.
}
