% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCountsArray.R
\name{subset.array}
\alias{subset.array}
\title{Subset an Array in one Dimension}
\usage{
\method{subset}{array}(x, dim, i, drop = FALSE)
}
\arguments{
\item{x}{an array.}

\item{dim}{an integer specifying the dimension to subset.}

\item{i}{subset index, see \code{\link{[}}.}

\item{drop}{logical indicating if singular dimensions should be dropped,
see \code{\link{[}}.}
}
\description{
Subset an Array in one Dimension
}
\examples{
hhh4contacts:::subset.array(counts, 1, 4:7)
}
\seealso{
the more general function \code{extract.array} in package
\pkg{R.utils}
}
\keyword{internal}
